refactor(types): выносит интерфейсы в отдельный файл
Перенесены интерфейсы в папку types. Названия типов,
используемых только в компонентах, заменены на Props.
