+++ src/components/offer/offer-inside/offer-inside.tsx
@@ -1,8 +1,8 @@
-interface OfferInsideProps {
+interface Props {
   goods: string[];
 }
 
-function OfferInside({ goods }: OfferInsideProps) {
+function OfferInside({ goods }: Props) {
   return (
     <div className="offer__inside">
       <h2 className="offer__inside-title">What&apos;s inside</h2>
