+++ src/mocks/favorite-places.ts
@@ -1,17 +1,4 @@
-export interface Place {
-  id: number;
-  title: string;
-  type: string;
-  price: number;
-  rating: number;
-  image: string;
-  isPremium: boolean;
-}
-
-export interface City {
-  city: string;
-  places: Place[];
-}
+import { City } from '../types/place';
 
 export const favoritePlaces: City[] = [
   {
