+++ src/types/place.ts
@@ -1,0 +1,14 @@
+export interface Place {
+  id: number;
+  title: string;
+  type: string;
+  price: number;
+  rating: number;
+  image: string;
+  isPremium: boolean;
+}
+
+export interface City {
+  city: string;
+  places: Place[];
+}
